backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pywebview==5.3

# Strands Agents
strands-agents>=0.2.0

# HTTP Client for OpenRouter
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
beautifulsoup4>=4.12.0
